Rewrite SMG: Change sizeOf return type used in expression evaluation from long to BigInteger
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38487 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/smg2/util/value/AddressEvaluator.java
@@ -8,6 +8,7 @@
 
 package org.sosy_lab.cpachecker.cpa.smg2.util.value;
 
+import java.math.BigInteger;
 import java.util.Collection;
 import org.sosy_lab.cpachecker.cfa.ast.c.CExpression;
 import org.sosy_lab.cpachecker.cpa.smg2.SMGState;
@@ -36,5 +37,5 @@ public interface AddressEvaluator {
   public CValueAndSMGState
       readValue(SMGState pSmgState, SMGObject pVariableObject, CExpression pIdExpression);
 
-  public long getBitSizeof(SMGState pInitialSmgState, CExpression pUnaryOperand);
+  public BigInteger getBitSizeof(SMGState pInitialSmgState, CExpression pUnaryOperand);
 }
